03302023-Update
Fixed issue with scriptable object not loading data to TilmapGridNode class. Used a data holder class called NodeProperties to be able to serialize node list.

Nodes are now found through a dictionary instead of a  bi dimensional array, this allows us to input direct world coordinates and not having to make a transformation every time we need to get a node.