04252023-Update
To prevent null exception with the Level Grid scriptable objects, UpdateTilemapProperties is called every time a scene is loaded.
Solved the issue with the automatic switching between scenes (Collision layer on barracks had a Dynamic rigid body so the collisions were falling on top of the exit)
Consider the event when a scene is loaded from the entrance triggered to the scene manager , needs to update cinemachine reference on BaseManager , player position and UI stats. Consider sending data to be persisted to a Game Manager class that will write to a file and reload when exiting battle scenes, All the objects that are persisted in the base should be destroyed during battle and reconstitued when the player returns to base