05052023-Update
Reviewed input controllers for the UI and player systems. Is likely that we we will have to map different control schemas for the battle and the battle UI as well . Revision in progress of both the UI and the Player input controllers. Also , some technical debt , remove the boolean from the playerControllerEnabled