06172023-Update
Solved Issue with tilemap loading , the way to go was to modify the exit condition in the UpdateTilemapProperties of the TilemapProperties class. The new exit condition requires that the tilemap layer count is only layer (to take into account the exterior maps) or if the count of nodeProperties list is already above 1  (meaning the list has been populated out of runtime) then the method is exited. This prevent doubly populating the Node dictionatry. This is not a very elegant solution but it seems to be able to read the information from multiple tilemap layers , while simultaneously allowing a tilemap to be populated outside of runtime