20230630-Update
Progress has been made in the valid movement list algorithm. However something very strange is happening in that the instantiated player units are not being assigned their gridPosition at runtime effectively. The units seem to be located by default at 0,0 and also they do not have a node assigned to them  so they are not tethered to the grid