20230710-Update
Little progress today after returning from vacation. Recursive call not working and is crashing the program for any movement range beyond 3 units. The cause for this is not yet determined but is possible that the while loop is somehow not reaching termination after a recursive call? is also possible that the program is crashing by not knowing how to handle positions outside the grid . The algorithm is not adequately finding the movement positions so it may be inadequate anyway. A solution could be to return to the original nested loop approach and reduce the movement range progressively as the positions are looked for or to modify it so no recursive calls are made.
The attack positions also do not make sense so that validation needs to be reviewed as well.