20230716-Update
Connected Action Menu  to move action and executed succesfully. However , there seems to be a number of bugs , notably , how it apparently is reading the edges of the player movement range as being outside of the movement range (likely because of the  transformation of coordinates between the selector transform and the actual grid coordinate)
Also of importance is the amount of spaghetti code we're getting from the interaction between unit and unit action scripts and the multiple events that have to occur for an action to be executed. The design of this code should be taken a look at in a more thorough way.