20230717-Update
Worked on updating the UnitActionSystem. One of the main problems is to have the menu pop up and dissapear when the action is called .
I solved the issues when the unit was deselecting and not moving to the edges of the movement range.
Still need to think on how to cleanup the code , reset unit status when movement is completed and the action is finished. Also , need to think on how to store all the unit valid movement positions for all actions (or would it be better to have attack, move and capture positions separate)